StorageManagement & IDE Added to Kernel. - Other stuff. - Things.
git-svn-id: https://spexeah.com:8443/svn/Asuro@457 6dbc8c32-bb84-406f-8558-d1cf31a0ab0c

+++ src/driver/net/ipv4.pas
@@ -60,7 +60,7 @@ end;
 
 procedure register;
 var
-    i : integer;
+    i : uint8;
 
 begin
     if not Registered then begin
